pyproject.toml
[project]
name = "mikrotik-routeros-mcp"
version = "0.1.0"
description = "Public MCP server for managing multiple MikroTik RouterOS devices with transport fallback and guarded write workflows."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "mcp",
  "paramiko>=3.4.0",
  "PyYAML>=6.0.2",
  "routeros-api>=0.21.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mikrotik_routeros_mcp"]

[project.scripts]
mikrotik-routeros-mcp = "mikrotik_routeros_mcp.server:main"

[dependency-groups]
dev = [
  "pytest>=9.0.2",
]
